docs and environment updates
