+++ onboardings/testhelper/testhelper.go
@@ -5,11 +5,7 @@ import (
 	"fmt"
 	"github.com/google/uuid"
 	"github.com/temporalio/temporal-jumpstart-golang/onboardings/domain/workflows"
-	"go.temporal.io/api/enums/v1"
-	"go.temporal.io/api/history/v1"
 	"go.temporal.io/sdk/client"
-	"google.golang.org/protobuf/encoding/protojson"
-	"os"
 	"reflect"
 )
 
@@ -22,45 +18,6 @@ func RandomString() string {
 
 // GetFunctionName shamelessly lifted from sdk-go
 var GetFunctionName = workflows.GetFunctionName
-
-// GetWorkflowHistory is a utility for fetching all history events out of a workflow execution
-func GetWorkflowHistory(ctx context.Context,
-	client client.Client,
-	workflowID string) (*history.History, error) {
-	result := &history.History{
-		Events: []*history.HistoryEvent{},
-	}
-	iter := client.GetWorkflowHistory(ctx, workflowID, "", true, enums.HISTORY_EVENT_FILTER_TYPE_ALL_EVENT)
-	for iter.HasNext() {
-		event, err := iter.Next()
-		if err != nil {
-			return nil, err
-		}
-		result.Events = append(result.Events, event)
-	}
-	return result, nil
-}
-
-// DumpWorkflowHistory is a utility for dumping all history events out of a workflow execution
-// Note that destinationPath is relative to any test's directory that is being run.
-func DumpWorkflowHistory(ctx context.Context,
-	client client.Client,
-	workflowID string,
-	destinationPath string) (*history.History, error) {
-	hist, err := GetWorkflowHistory(ctx, client, workflowID)
-	if err != nil {
-		return nil, err
-	}
-	data, err := protojson.Marshal(hist)
-	if err != nil {
-		return nil, err
-	}
-	err = os.WriteFile(destinationPath, data, os.ModePerm)
-	if err != nil {
-		return nil, err
-	}
-	return hist, nil
-}
 
 // TestEncodedValue simplifies testing with this result from the Temporal Client
 type TestEncodedValue struct {
@@ -83,6 +40,7 @@ func (val *TestEncodedValue) Get(valuePtr interface{}) error {
 	return nil
 }
 
+// TestWorkflowRun implements WorkflowRun for testing
 type TestWorkflowRun struct {
 	WorkflowID string
 	RunID      string
@@ -108,6 +66,7 @@ func (t *TestWorkflowRun) GetWithOptions(ctx context.Context, valuePtr interface
 	panic("implement me")
 }
 
+// TestWorkflowUpdateHandler implements WorkflowUpdateHandle for testing
 type TestWorkflowUpdateHandle struct {
 	WorkflowIDToUse string
 }
